Continue separation of FunctionalTestCase from TwillTestCase.
Move non-twill stuff out into its own module including FunctionalTestCase.
